// front/src/app/models/bird.ts
export class Bird {
    _id: string; 
    latinski_naziv: string;
    naziv: string;
    drugi_naziv: string;
    staniste: string;
    porodica: string;
    slika_m: string;
    slika_z: string;
    tip: string;
    boja_m: string;
    boja_z: string;
    opis: string;
    duzina_tela: string;
    raspon_krila: string;
    tezina: string;
    ishrana: string;
    gnezdenje: string;
    prisutnost: string;
    snimak: string;
    pesma: string;
    zov_zenke: string;
    zov: string;
    zov_u_letu: string;
    uzbuna: string;
    zov_mladunaca: string;
    br: number;
    engleski_naziv: string;
  }
  